ttl: merge colinfo/ttl.go into catpb/ttl.go to avoid a circular dependency
Release note: None
