+++ pkg/sql/alter_table.go
@@ -25,7 +25,6 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog"
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/catalogkeys"
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/catpb"
-	"github.com/cockroachdb/cockroach/pkg/sql/catalog/colinfo"
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/descpb"
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/funcdesc"
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/resolver"
@@ -489,7 +488,7 @@ func (n *alterTableNode) startExec(params runParams) error {
 			}
 
 			tableDesc := n.tableDesc
-			if t.Column == colinfo.TTLDefaultExpirationColumnName &&
+			if t.Column == catpb.TTLDefaultExpirationColumnName &&
 				tableDesc.HasRowLevelTTL() &&
 				tableDesc.GetRowLevelTTL().HasDurationExpr() {
 				return errors.WithHintf(
@@ -601,7 +600,7 @@ func (n *alterTableNode) startExec(params runParams) error {
 					"column %q in the middle of being dropped", t.GetColumn())
 			}
 			columnName := col.GetName()
-			if columnName == colinfo.TTLDefaultExpirationColumnName &&
+			if columnName == catpb.TTLDefaultExpirationColumnName &&
 				tableDesc.HasRowLevelTTL() &&
 				tableDesc.GetRowLevelTTL().HasDurationExpr() {
 				return pgerror.Newf(
@@ -747,7 +746,7 @@ func (n *alterTableNode) startExec(params runParams) error {
 		case *tree.AlterTableRenameColumn:
 			tableDesc := n.tableDesc
 			columnName := t.Column
-			if columnName == colinfo.TTLDefaultExpirationColumnName &&
+			if columnName == catpb.TTLDefaultExpirationColumnName &&
 				tableDesc.HasRowLevelTTL() &&
 				tableDesc.GetRowLevelTTL().HasDurationExpr() {
 				return pgerror.Newf(
@@ -1835,7 +1834,7 @@ func handleTTLStorageParamChange(
 
 		// Update default expression on automated column if required.
 		if before.HasDurationExpr() && after.HasDurationExpr() && before.DurationExpr != after.DurationExpr {
-			col, err := catalog.MustFindColumnByName(tableDesc, colinfo.TTLDefaultExpirationColumnName)
+			col, err := catalog.MustFindColumnByName(tableDesc, catpb.TTLDefaultExpirationColumnName)
 			if err != nil {
 				return false, err
 			}
@@ -1877,11 +1876,11 @@ func handleTTLStorageParamChange(
 		// Adding a TTL requires adding the automatic column and deferring the TTL
 		// addition to after the column is successfully added.
 		addTTLMutation = true
-		if catalog.FindColumnByName(tableDesc, colinfo.TTLDefaultExpirationColumnName) != nil {
+		if catalog.FindColumnByName(tableDesc, catpb.TTLDefaultExpirationColumnName) != nil {
 			return false, pgerror.Newf(
 				pgcode.InvalidTableDefinition,
 				"cannot add TTL to table with the %s column already defined",
-				colinfo.TTLDefaultExpirationColumnName,
+				catpb.TTLDefaultExpirationColumnName,
 			)
 		}
 		col, err := rowLevelTTLAutomaticColumnDef(after)
@@ -1927,7 +1926,7 @@ func handleTTLStorageParamChange(
 		// Create the DROP COLUMN job and the associated mutation.
 		dropTTLMutation = true
 		droppedViews, err := dropColumnImpl(params, tn, tableDesc, after, &tree.AlterTableDropColumn{
-			Column: colinfo.TTLDefaultExpirationColumnName,
+			Column: catpb.TTLDefaultExpirationColumnName,
 		})
 		if err != nil {
 			return false, err
