+++ pkg/sql/catalog/catpb/BUILD.bazel
@@ -41,7 +41,6 @@ go_library(
         "job_id.go",
         "multiregion.go",
         "privilege.go",
-        "ttl.go",
         ":gen-privilegedescversion-stringer",  # keep
     ],
     embed = [":catpb_go_proto"],
