--- pkg/sql/catalog/catpb/ttl.go
@@ -1,19 +1,0 @@
-// Copyright 2022 The Cockroach Authors.
-//
-// Use of this software is governed by the Business Source License
-// included in the file licenses/BSL.txt.
-//
-// As of the Change Date specified in that file, in accordance with
-// the Business Source License, use of this software will be governed
-// by the Apache License, Version 2.0, included in the file
-// licenses/APL.txt.
-
-package catpb
-
-// DeletionCronOrDefault returns the DeletionCron or the global default.
-func (m *RowLevelTTL) DeletionCronOrDefault() string {
-	if override := m.DeletionCron; override != "" {
-		return override
-	}
-	return "@hourly"
-}
