+++ pkg/sql/catalog/tabledesc/ttl.go
@@ -16,7 +16,6 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog"
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/catenumpb"
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/catpb"
-	"github.com/cockroachdb/cockroach/pkg/sql/catalog/colinfo"
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/schemaexpr"
 	"github.com/cockroachdb/cockroach/pkg/sql/parser"
 	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgcode"
@@ -102,16 +101,16 @@ func ValidateTTLExpirationColumn(desc catalog.TableDescriptor) error {
 		return nil
 	}
 	intervalExpr := desc.GetRowLevelTTL().DurationExpr
-	col, err := catalog.MustFindColumnByTreeName(desc, colinfo.TTLDefaultExpirationColumnName)
+	col, err := catalog.MustFindColumnByTreeName(desc, catpb.TTLDefaultExpirationColumnName)
 	if err != nil {
-		return errors.Wrapf(err, "expected column %s", colinfo.TTLDefaultExpirationColumnName)
+		return errors.Wrapf(err, "expected column %s", catpb.TTLDefaultExpirationColumnName)
 	}
 	expectedStr := `current_timestamp():::TIMESTAMPTZ + ` + string(intervalExpr)
 	if col.GetDefaultExpr() != expectedStr {
 		return pgerror.Newf(
 			pgcode.InvalidTableDefinition,
 			"expected DEFAULT expression of %s to be %s",
-			colinfo.TTLDefaultExpirationColumnName,
+			catpb.TTLDefaultExpirationColumnName,
 			expectedStr,
 		)
 	}
@@ -119,7 +118,7 @@ func ValidateTTLExpirationColumn(desc catalog.TableDescriptor) error {
 		return pgerror.Newf(
 			pgcode.InvalidTableDefinition,
 			"expected ON UPDATE expression of %s to be %s",
-			colinfo.TTLDefaultExpirationColumnName,
+			catpb.TTLDefaultExpirationColumnName,
 			expectedStr,
 		)
 	}
