+++ pkg/sql/ttl/ttljob/ttljob.go
@@ -25,7 +25,6 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/settings/cluster"
 	"github.com/cockroachdb/cockroach/pkg/sql"
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/catpb"
-	"github.com/cockroachdb/cockroach/pkg/sql/catalog/colinfo"
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/descs"
 	"github.com/cockroachdb/cockroach/pkg/sql/execinfrapb"
 	"github.com/cockroachdb/cockroach/pkg/sql/isql"
@@ -150,7 +149,7 @@ func (t rowLevelTTLResumer) Resume(ctx context.Context, execCtx interface{}) err
 		return err
 	}
 
-	ttlExpr := colinfo.DefaultTTLExpirationExpr
+	ttlExpr := catpb.DefaultTTLExpirationExpr
 	if rowLevelTTL.HasExpirationExpr() {
 		ttlExpr = "(" + rowLevelTTL.ExpirationExpr + ")"
 	}
